WIP: move player into custom_user
