+++ backend/game/models.py
@@ -1,4 +1,5 @@
 # for displaying games in admin panel
+from custom_user.models import Player
 from django.contrib import admin
 from django.db import models
 
@@ -9,8 +10,8 @@ from django.utils import timezone
 # create game when starting a new game
 # update game when finishing a game/goals are scored
 class Game(models.Model):
-	player1 = models.ForeignKey('Player', related_name='games_as_player1', on_delete=models.CASCADE)
-	player2 = models.ForeignKey('Player', related_name='games_as_player2', on_delete=models.CASCADE)
+	player1 = models.ForeignKey(Player, related_name='games_as_player1', on_delete=models.CASCADE)
+	player2 = models.ForeignKey(Player, related_name='games_as_player2', on_delete=models.CASCADE)
 	score1 = models.IntegerField(default=0)
 	score2 = models.IntegerField(default=0)
 	# calculate duration of game from start to finish
@@ -32,33 +33,3 @@ class Game(models.Model):
 	)
 	def __str__(self):
 		return f'{self.player1} vs {self.player2} ({self.score1}-{self.score2})'
-
-
-class Player(models.Model):
-	matches_won = models.IntegerField(default=0)
-	matches_lost = models.IntegerField(default=0)
-	user = models.OneToOneField(
-		'custom_user.CustomUser',
-		on_delete=models.CASCADE,
-		related_name='player_user',
-		null=True,
-		blank=True,
-	)
-
-	# The statistics for the quiz game
-	quiz_games_played = models.IntegerField(default=0)
-	quiz_games_won = models.IntegerField(default=0)
-	quiz_total_score = models.BigIntegerField(default=0)
-	quiz_high_score = models.IntegerField(default=0)
-	quiz_questions_asked = models.IntegerField(default=0)
-	quiz_correct_answers = models.IntegerField(default=0)
-
-	def __str__(self):
-		return self.user.username
-
-	def win_to_loss_ratio(self):
-		if self.matches_lost == 0:
-			return self.matches_won
-		if self.matches_lost == 0:
-			return self.matches_won
-		return round(self.matches_won / self.matches_lost, 2)
